requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
azure-core==1.39.0
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
great_expectations==1.15.2
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
marshmallow==3.26.2
mistune==3.2.0
narwhals==2.18.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
py4j==0.10.9.7
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pyspark==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
ruamel.yaml==0.19.1
scipy==1.17.1
six==1.17.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
